[flakes] reset nix profile directory when Flakes feature is toggled
