+++ internal/nix/nix.go
@@ -19,7 +19,8 @@ import (
 )
 
 // ProfilePath contains the contents of the profile generated via `nix-env --profile ProfilePath <command>`
-// Instead of using directory, prefer using the devbox.ProfilePath() function that ensures the directory exists.
+// or `nix profile install --profile ProfilePath <package...>`
+// Instead of using directory, prefer using the devbox.ProfileDir() function that ensures the directory exists.
 const ProfilePath = ".devbox/nix/profile/default"
 
 var ErrPackageNotFound = errors.New("package not found")
